Added tower money deduction and increased upgrade collider box

+++ Assets/Scripts/GameController.cs
@@ -194,9 +194,9 @@ public class GameController : MonoBehaviour
         money_text.text = money.ToString();
         updateTowerAvailability(); 
     }
-    public float moneyAmt()
+    public int moneyAmt()
     {
-        return money; 
+        return (int)money; 
     }
     public void collapseUINotPicking()
     {
